requirements.txt
# Django core and API
dj-database-url==3.0.1
Django==5.0.14
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
asgiref==3.9.1
pillow==11.3.0
platformdirs==4.3.8
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.9.0
python-decouple==3.8
python-dotenv==1.0.1

# Environment and configuration
python-dotenv

# Database drivers
mysqlclient
psycopg2-binary

# Image handling
pillow

# HTTP requests
requests

# Deployment tools
gunicorn
whitenoise

# API documentation (optional but useful)
drf-yasg
coreapi
